+++ packages/admin-api-client/index.js
@@ -1,2 +1,1 @@
-export { configClient } from "./src/entities";
 export * from "./src/methods";