+++ packages/cypress-commands/test/commands.spec.js
@@ -11,6 +11,26 @@ const apiClient = {
   configClient: doNothing,
 };
 
+class FakeAdminApiClient {
+  updateConfig(config) {
+    apiClient.updateConfig(config);
+  }
+
+  useRouteVariant(id) {
+    apiClient.useRouteVariant(id);
+  }
+
+  restoreRouteVariants() {
+    apiClient.restoreRouteVariants();
+  }
+
+  configClient(config) {
+    apiClient.configClient(config);
+  }
+}
+
+apiClient.AdminApiClient = FakeAdminApiClient;
+
 jest.mock("@mocks-server/admin-api-client", () => {
   return apiClient;
 });
@@ -22,7 +42,6 @@ import { commands } from "../src/commands";
 describe("commands", () => {
   let sandbox;
   let cypressMock;
-  let setCollection, setDelay, setConfig, useRouteVariant, restoreRouteVariants, configClient;
 
   beforeEach(() => {
     sandbox = sinon.createSandbox();
@@ -31,8 +50,6 @@ describe("commands", () => {
     sandbox.stub(apiClient, "restoreRouteVariants");
     sandbox.stub(apiClient, "configClient");
     cypressMock = new CypressMock();
-    ({ setCollection, setDelay, setConfig, useRouteVariant, restoreRouteVariants, configClient } =
-      commands(cypressMock.stubs));
   });
 
   afterEach(() => {
@@ -40,22 +57,14 @@ describe("commands", () => {
   });
 
   describe("when initializing", () => {
-    it("should call to set port config if env var is defined", () => {
-      cypressMock.stubs.env.withArgs("MOCKS_SERVER_ADMIN_API_PORT").returns("foo");
+    it("should call to set client config", () => {
+      cypressMock.stubs.env.withArgs("MOCKS_SERVER_ADMIN_API_PORT").returns("foo-port");
+      cypressMock.stubs.env.withArgs("MOCKS_SERVER_ADMIN_API_HOST").returns("foo-host");
       commands(cypressMock.stubs);
       expect(
         apiClient.configClient.calledWith({
-          port: "foo",
-        })
-      ).toBe(true);
-    });
-
-    it("should call to set baseUrl config if env var is defined", () => {
-      cypressMock.stubs.env.withArgs("MOCKS_SERVER_ADMIN_API_HOST").returns("foo");
-      commands(cypressMock.stubs);
-      expect(
-        apiClient.configClient.calledWith({
-          host: "foo",
+          port: "foo-port",
+          host: "foo-host",
         })
       ).toBe(true);
     });
@@ -63,6 +72,7 @@ describe("commands", () => {
 
   describe("setCollection command", () => {
     it("should call to set current collection", () => {
+      const { setCollection } = commands(cypressMock.stubs);
       setCollection("foo");
       expect(
         apiClient.updateConfig.calledWith({
@@ -77,6 +87,7 @@ describe("commands", () => {
 
     it("should do nothing if plugin is disabled", () => {
       cypressMock.stubs.env.returns("false");
+      const { setCollection } = commands(cypressMock.stubs);
       setCollection("foo");
       expect(apiClient.updateConfig.callCount).toEqual(0);
     });
@@ -84,6 +95,7 @@ describe("commands", () => {
 
   describe("setDelay command", () => {
     it("should call to update delay", () => {
+      const { setDelay } = commands(cypressMock.stubs);
       setDelay(3000);
       expect(
         apiClient.updateConfig.calledWith({
@@ -98,6 +110,7 @@ describe("commands", () => {
 
     it("should do nothing if plugin is disabled", () => {
       cypressMock.stubs.env.returns("false");
+      const { setDelay } = commands(cypressMock.stubs);
       setDelay("foo");
       expect(apiClient.updateConfig.callCount).toEqual(0);
     });
@@ -105,12 +118,14 @@ describe("commands", () => {
 
   describe("setConfig command", () => {
     it("should call to update config", () => {
+      const { setConfig } = commands(cypressMock.stubs);
       setConfig("foo");
       expect(apiClient.updateConfig.calledWith("foo")).toBe(true);
     });
 
     it("should do nothing if plugin is disabled", () => {
       cypressMock.stubs.env.returns("0");
+      const { setConfig } = commands(cypressMock.stubs);
       setConfig("foo");
       expect(apiClient.updateConfig.callCount).toEqual(0);
     });
@@ -118,12 +133,14 @@ describe("commands", () => {
 
   describe("useRouteVariant command", () => {
     it("should call to useRoute variant", () => {
+      const { useRouteVariant } = commands(cypressMock.stubs);
       useRouteVariant("foo");
       expect(apiClient.useRouteVariant.calledWith("foo")).toBe(true);
     });
 
     it("should do nothing if plugin is disabled", () => {
       cypressMock.stubs.env.returns(0);
+      const { useRouteVariant } = commands(cypressMock.stubs);
       useRouteVariant("foo");
       expect(apiClient.useRouteVariant.callCount).toEqual(0);
     });
@@ -131,12 +148,14 @@ describe("commands", () => {
 
   describe("restoreRouteVariants command", () => {
     it("should call to useRoute variant", () => {
+      const { restoreRouteVariants } = commands(cypressMock.stubs);
       restoreRouteVariants();
       expect(apiClient.restoreRouteVariants.callCount).toEqual(1);
     });
 
     it("should do nothing if plugin is disabled", () => {
       cypressMock.stubs.env.returns(false);
+      const { restoreRouteVariants } = commands(cypressMock.stubs);
       restoreRouteVariants();
       expect(apiClient.restoreRouteVariants.callCount).toEqual(0);
     });
@@ -144,25 +163,29 @@ describe("commands", () => {
 
   describe("configClient method", () => {
     it("should call to config admin-api-client host", () => {
+      const { configClient } = commands(cypressMock.stubs);
       configClient({
         host: "foo",
       });
       expect(
         apiClient.configClient.calledWith({
           host: "foo",
+          port: undefined,
         })
       ).toBe(true);
     });
 
     it("should call to config admin-api-client apiPath port and host", () => {
+      const { configClient } = commands(cypressMock.stubs);
       configClient({
-        host: "foo",
-        port: "foo-2",
+        host: "foo-host",
+        port: "foo-port",
       });
+
       expect(
         apiClient.configClient.calledWith({
-          host: "foo",
-          port: "foo-2",
+          host: "foo-host",
+          port: "foo-port",
         })
       ).toBe(true);
     });